add db to project

+++ app/schemas.py
@@ -1,14 +1,16 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=255)
 
 
     @field_validator('amount')
-    def amount_must_be_positive(cls, value: float) -> float:
+    def amount_must_be_positive(cls, value: Decimal) -> Decimal:
         if value <= 0:
             raise ValueError(f"{cls.__name__} Amount must be positive")
         return value
@@ -24,7 +26,7 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=127)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator('name')
     def name_not_empty(cls, value: str) -> str:
@@ -34,7 +36,7 @@ class CreateWalletRequest(BaseModel):
         return value
 
     @field_validator('initial_balance')
-    def balance_not_negative(cls, value: float) -> float:
+    def balance_not_negative(cls, value: Decimal) -> Decimal:
         if value < 0:
             raise ValueError(f"{cls.__name__} Balance cannot be negative")
         return value
